Exercise 7: More Printing
from: http://learnrubythehardway.org/book/ex7.html
Now we are going to do a bunch of exercises where you just type code in and make it run. I won't be explaining this exercise because it is more of the same. The purpose is to build up your chops. See you in a few exercises, and do not skip! Do not paste!